Add files via upload
Project Title: ATM Management System using C++

ATMs are Automated Teller Machines that are used to carry day-to-day financial transactions. ATMs can be used to withdraw money or to deposit money or even to know the information of an account like the balance amount, etc. They are convenient and easy to use,and  it allows consumers to perform quick self-service transactions.

In this article, we will discuss the ATM Management System in C++ which is an application that provides users with every aspect that an actual Automated Teller Machine i.e., ATM should have. It is a menu-driven program having ATM functions which include:

1. Enter Name, Account number, Account type to be shown during transactions.
2. Shows the information about the person who is doing the transaction.
3. Enter amount to deposited in the account.
4. Shows the Balance in the account.
5. Enter amount to be withdrawn from the account, and then it shows available balance.
6. Cancel the transaction.